[BasicPrunning] First commit of basic Pruning.
The implementation is based on:
- Each time an mpt node is inserted or deleted, the references it
has are increased ir decreased respectively
- When a node reaches 0 references, it's tagged as to be removed
candidate within that block
- Everytime we update the bestBlockNumber we check X blocks back
and remove all nodes with 0 references.
- Genesis and Mining blocks are not yet cleaned

Things to improve:
- Pruning should be moved to other plase than where the best block
number is updated. This was made in order to make it easier to
synchronize concurrent reads and pruning but it's not the best idea
- Genesis is a fixed amount of data and we can live without pruning
it but we should take care of mined blocks