Proof of key membership in the MPT
Given a key and assuming the root hash is known, we provide a proof that the
key exists in the Merkle Patricia Tree.

We also verify the membership proof iff the key exists and expose the proof
functionality via an extension of the current JSON-RPC protocol.

The current implementation of verification is not meant as a complete one;
rather it is a first, non-general approximation that has been developed to ease
testing of the proof algorithm.

Fundamentally, proof and verification are the same algorithm that walks a tree,
the difference being in the action done at each node. Proof accumulates node
hashes, verification checks those (previously accumulated) hashes. So, in true
modular style, a unified proof/verification algorithm is a composition of tree
traversal plus the action. The ingredients are present, so an implementation
should not be far from the current state (check `walk()` method inside
`MerklePatriciaTrie.prove()`).