Selective forging
`mining-enabled` has been silently ignored by the code in the case
of `atomix-raft` consensus. The code just assumes a value of
`true`. The assumption is correct regarding the current design,
since all nodes participate in the Raft consensus, but we need
more operational flexibility. In particular, we may need to
temporarily disable forging until a (recovering) node completely
syncs the blockchain and re-enable it afterwards, thus resuming
normal operation.

The current patch inspects the value of `mining-enabled` and acts
accordingly.

Ref CGKIELE-318