Export build-related info
From a running node, it is now possible to know exactly which
commit the code is based on, along with some other metadata.

This _buildinfo_ is exported via the logging subsystem when the
node boots, and it is also available under the `/buildinfo`
HTTP endpoint (using the HTTP mode of JSON/RPC).

Ref CGKIELE-463